+++ changelog.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Upgraded to all latest CI
+- Upgraded to all latest CI
 
 ## [1.0.0] - 2024-06-13
 
@@ -34,5 +35,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]: https://github.com/ortus-boxlang/bx-esapi/compare/v1.1.0...HEAD
 
 [1.1.0]: https://github.com/ortus-boxlang/bx-esapi/compare/v1.0.0...v1.1.0
+
 
 [1.0.0]: https://github.com/ortus-boxlang/bx-esapi/compare/251f3772e721f1f7aea3f7d2e2da602b8af97a40...v1.0.0